refactor display
