+++ src/main/kotlin/org/jetbrains/research/testspark/bundles/BundlePaths.kt
@@ -1,10 +1,8 @@
 package org.jetbrains.research.testspark.bundles
 
-class BundlePaths {
-    companion object {
-        const val DEFAULTS: String = "defaults.TestSpark"
-        const val MESSAGES: String = "messages.TestSpark"
-        const val LABELS: String = "defaults.Labels"
-        const val TOOLTIPS: String = "defaults.Tooltips"
-    }
+object BundlePaths {
+    const val DEFAULTS: String = "defaults.TestSpark"
+    const val MESSAGES: String = "messages.TestSpark"
+    const val LABELS: String = "defaults.Labels"
+    const val TOOLTIPS: String = "defaults.Tooltips"
 }
