+++ src/main/kotlin/org/jetbrains/research/testspark/display/common/TestCaseDocumentCreator.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.testspark.display
+package org.jetbrains.research.testspark.display.common
 
 import com.intellij.lang.Language
 import com.intellij.openapi.application.ApplicationManager
